Get Quote response completed

+++ backend/services/OrderService/src/main/java/com/swifttrack/OrderService/dto/ModelQuoteInput.java
@@ -1,10 +1,10 @@
 package com.swifttrack.OrderService.dto;
 
 public record ModelQuoteInput(
-        String provider,
-        double distance_km,
-        double traffic_level,
-        boolean is_peak_hour,
-        double provider_load) {
+                String provider,
+                double distance_km,
+                double traffic_level,
+                boolean is_peak_hour,
+                int provider_load) {
 
 }
